Arenex/minecraft-runner/requirements.txt
httpx==0.27.0
asyncio
python-dotenv
rich
redis
